backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
groq==1.1.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.6.0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
